Improve documentation for serde_as types and traits
* Add documentation to the `de` and `ser` modules.
* Expose the types `SerializeAsWrap` and `DeserializeAsWrap`. These
  types, or similar ones, are necessary to implement `SerializeAs` for
  container types, as the wrapper allows to change how the inner types
  get serialized.
* Add documentation to the traits `SerializeAs` and `DeserializeAs`.
